Finally, all tests pass.
On to benchmarking and cleaning it up, if the results are good.

+++ src/main/scala/com/signalcollect/triplerush/PlaceholderEdge.scala
@@ -25,7 +25,10 @@ import com.signalcollect.interfaces.EdgeId
 import com.signalcollect.GraphEditor
 import com.signalcollect.Vertex
 
-class PlaceholderEdge(val childDelta: Int) extends Edge[Int] {
+/**
+ * Trickery! It's actually the child delta (an Int), but we just pretend it's a Long.
+ */
+class PlaceholderEdge(val childDelta: Int) extends Edge[Long] {
   override def id: EdgeId[_] = throw new UnsupportedOperationException
   override def sourceId: Any = throw new UnsupportedOperationException
   override def targetId = childDelta
